Remove matrix * vector specialization for strided arrays
This method entirely duplicates the one below it with the exception that
it also does a `convert` on one of the inputs with the result of
`promote_op`, which doesn't make a whole lot of sense in some cases. By
virtue of calling `mul!`, strided arrays that go through the abstract
array method will still hit the same optimized methods that use BLAS.

Fixes #32092.