effects: add `:noglobal` effect
This is a preparatory PR for future improvements on the effects analysis.
This commit adds the `:noglobal` helper effect, that tracks if a method
involves any access or modification on any mutable global state. This
effect is actually a subset of the existing `:effect_free` effect, and
it can be considered as a very limited escape analysis that tracks
escapability of local mutable objects. This more fine-grained
`:effect_free` information will allow us to prove `:consistent`-cy or
`:effect_free`-ness even in a presence of allocations or modifications
of local mutable allocations in some cases when the `:noglobal` effect
has been proven. Separate PRs that actually improve the effect analysis
accuracy will follow.